requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
matplotlib>=3.7.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
schedule>=1.2.0
